+++ cmp/grid/GridModel.js
@@ -129,6 +129,8 @@ export class GridModel extends HoistModel {
     @observable.ref columnState = [];
     /** @member {Object} */
     @observable.ref expandState = {};
+    /** @member {AutosizeState} */
+    @observable.ref autosizeState = {};
     /** @member {GridSorter[]} */
     @observable.ref sortBy = [];
     /** @member {string[]} */
@@ -768,7 +770,10 @@ export class GridModel extends HoistModel {
 
         this.columns = columns;
         this.columnState = this.getLeafColumns()
-            .map(({colId, width, hidden, pinned}) => ({colId, width, hidden, pinned}));
+            .map(it => {
+                const {colId, width, hidden, pinned} = it;
+                return {colId, width, hidden, pinned};
+            });
     }
 
     /** @param {ColumnState[]} colState */
@@ -824,6 +829,24 @@ export class GridModel extends HoistModel {
         }
     }
 
+    @action
+    setAutosizeState(autosizeState) {
+        if (!equal(this.autosizeState, autosizeState)) {
+            this.autosizeState = deepFreeze(autosizeState);
+        }
+    }
+
+    noteColumnsManuallySized(colIds) {
+        const colStateChanges = castArray(colIds).map(colId => ({colId, manuallySized: true}));
+        this.applyColumnStateChanges(colStateChanges);
+    }
+
+    noteColumnsAutosized(colIds) {
+        const colStateChanges = castArray(colIds).map(colId => ({colId, manuallySized: false}));
+        this.applyColumnStateChanges(colStateChanges);
+        this.setAutosizeState({sizingMode: this.sizingMode});
+    }
+
     /**
      * This method will update the current column definition if it has changed.
      * Throws an exception if any of the columns provided in colStateChanges are not
@@ -852,6 +875,7 @@ export class GridModel extends HoistModel {
             if (!isNil(change.width)) col.width = change.width;
             if (!isNil(change.hidden)) col.hidden = change.hidden;
             if (!isUndefined(change.pinned)) col.pinned = change.pinned;
+            if (!isNil(change.manuallySized)) col.manuallySized = change.manuallySized;
         });
 
         // 2) If the changes provided is a full list of leaf columns, synchronize the sort order
@@ -1038,7 +1062,6 @@ export class GridModel extends HoistModel {
             .linkTo(this.autosizeTask);
     }
 
-
     /**
      * Begin an inline editing session.
      * @param {RecordOrId} [recOrId] - Record/ID to edit. If unspecified, the first selected Record
@@ -1178,6 +1201,7 @@ export class GridModel extends HoistModel {
 
         try {
             await XH.gridAutosizeService.autosizeAsync(this, colIds, options);
+            this.noteColumnsAutosized(colIds);
         } finally {
             if (showMask) {
                 await wait();
@@ -1307,10 +1331,10 @@ export class GridModel extends HoistModel {
     // conflict with any code-level updates to their widths.
     removeTransientWidths(columnState) {
         const gridCols = this.getLeafColumns();
-
         return columnState.map(state => {
             const col = this.findColumn(gridCols, state.colId);
-            return col.resizable ? state : omit(state, 'width');
+            if (!col.resizable) state = omit(state, 'width');
+            return state;
         });
     }
 
@@ -1506,6 +1530,12 @@ export class GridModel extends HoistModel {
  * @property {number} [width] - new width to set for the column
  * @property {boolean} [hidden] - visibility of the column
  * @property {string} [pinned] - 'left'|'right' if pinned, null if not
+ * @property {boolean} [manuallySized] - has this column been resized manually?
+ */
+
+/**
+ * @typedef {Object} AutosizeState
+ * @property {SizingMode} sizingMode - sizing mode used last time the columns were autosized.
  */
 
 /**